# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi untuk Backend API Point of Sale (POS) MVP.",
        "title": "POS MVP API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support Tim",
            "email": "support@posmvp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/transactions": {
            "get": {
                "description": "Mengambil semua daftar transaksi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get All Transaksi",
                "parameters": [
                    {
                        "description": "Kredensial Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil Get All Transaction",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi/Input Invalid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial Tidak Valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Membuat transaksi baru.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Create Transaksi",
                "parameters": [
                    {
                        "description": "Kredensial Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil Get Transaction",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi/Input Invalid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial Tidak Valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/:id": {
            "get": {
                "description": "Mengambil 1 transaksi berdasarkan id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Transaksi",
                "parameters": [
                    {
                        "description": "Kredensial Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.TransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil Get Transaction",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi/Input Invalid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial Tidak Valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "services.ItemRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "services.TransactionRequest": {
            "type": "object",
            "required": [
                "cash",
                "items",
                "payment_method"
            ],
            "properties": {
                "cash": {
                    "description": "Uang yang dibayarkan pelanggan",
                    "type": "number",
                    "minimum": 0
                },
                "discount": {
                    "type": "number",
                    "minimum": 0
                },
                "items": {
                    "description": "Daftar produk yang dibeli",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/services.ItemRequest"
                    }
                },
                "payment_method": {
                    "description": "e.g., \"Cash\", \"QRIS\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}